enum-done

+++ Filter.cs
@@ -40,7 +40,7 @@ namespace StoreManager
             private string stringValue()
             {
                 if (value is Enum)
-                    return Convert.ToInt32(value).ToString();
+                    return Convert.ToInt16(value).ToString();
                 if (value.GetType() != typeof(DateTime))
                     return value.ToString();
                 return ((DateTime)value).ToString("yyyy-MM-dd HH:mm:ss");
